Creating the Voting card and summary pages
I have created a button for the engineers and team leader to start their health check voting stage where they get 12 cards to fill out, each new stage creates a session and is logged ontop the the database. whenever a department leader or ssenior manager logs in they have a view summaries button to view the summaries of the votes. All the votes are tallied by the state which they answered in the voting cards and after being taken from the database a pi chart for each team and department they belong to is created and is shown on the summaries page.